refactor: 라우터 변수명 통일, 파일명 API에 맞춰 통일
