+++ src/server.js
@@ -4,7 +4,7 @@ import express from "express";
 import { errorHandler } from "./middlewares/errorHandler.js";
 import adKeywordRoutes from "./routes/ad-keywords/index.js";
 import radioChannelRoutes from "./routes/radio-channels/index.js";
-import reportRouter from "./routes/report/index.js";
+import reportRoutes from "./routes/reports/index.js";
 import userRoutes from "./routes/users/index.js";
 
 dotenv.config();
@@ -14,9 +14,9 @@ const app = express();
 app.use(express.json());
 
 app.use("/ad-keywords", adKeywordRoutes);
-app.use("/radio-channels", radioChannelRoutes);
-app.use("/reports", reportRouter);
 app.use("/users", userRoutes);
+app.use("/reports", reportRoutes);
+app.use("/radio-channels", radioChannelRoutes);
 
 app.use(errorHandler);
